requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
cryptography>=41.0.0
pynacl>=1.5.0
python-dateutil>=2.8.0
jsonschema>=4.17.0
